os/board/rtl8730e: Sync to 8.3c SDK for TizenRT WiFi specific: -Bug fixes for WiFi stability -Enable WiFi high throughput -Adjust WiFi power by rate for Samsung Others: -Synchronise code for Peripheral, BLE, WiFi based on 8.3c SDK
SDK 8.3c : major update
Added -Werror option for the AP makefile
Upgrade lwip to version 2.1.2
Upgrade the audio feature
Upgrade and Fix bluetooth related feature
Replace the fatfs and littlefs example with the vfs example
Optimize wifi-audio layer
Modify OTA HTTP tool and Example
Simplify elfbin/imgtool.sh and unify rdpiv and rsip iv
Remove the media cache manager
Update pre-build lib
some WiFi related bug fix, new structure & ipc layer changes

+++ os/board/rtl8730e/src/component/audio/interfaces/hardware/audio/audio_hw_capture.h
@@ -65,7 +65,7 @@ struct AudioHwCapture {
 	 * @return Returns <b>0</b> if capture frames and timestamp get successfully;
 	 * returns a negative value otherwise.
 	 */
-	int (*GetCapturePosition)(const struct AudioHwCapture *stream, uint32_t *frames, struct timespec *timestamp);
+	int (*GetCapturePosition)(const struct AudioHwCapture *stream, uint64_t *frames, struct timespec *timestamp);
 
 	/**
 	 * @brief Get system latency of the current <b>AudioHwCapture</b> object .
